% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FScanR.R
\name{FScanR}
\alias{FScanR}
\title{FScanR}
\usage{
FScanR(
  blastx_output,
  mismatch_cutoff = 5,
  evalue_cutoff = 1e-05,
  frameDist_cutoff = 10
)
}
\arguments{
\item{blastx_output}{Input file with 14 columns in tab-delimited format, output from BLASTX using parameters: 
-outfmt '6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore qframe sframe'}

\item{mismatch_cutoff}{Threshold of number of mismatches for BLASTX hits, default 5 (aa)}

\item{evalue_cutoff}{Threshold of E-value for BLASTX hits, default 1e-5}

\item{frameDist_cutoff}{Threshold for gap size (bp) to detect frameshifting between BLASTX hits of same mRNA/cDNA sequence, default 10 (nt)}
}
\value{
dataframe
}
\description{
'FScanR' identifies Programmed Ribosomal Frameshifting (PRF) events from BLASTX homolog sequence alignment 
between targeted genomic/cDNA/mRNA sequences against the peptide library of the same species or a close relative.
}
\details{
The output by BLASTX or diamond BLASTX will be used as input of 'FScanR' and should be in a tabular format with 14 columns. 

For BLASTX, the output parameter should be: -outfmt '6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore qframe sframe'. 

For diamond BLASTX, the output parameter should be: -outfmt 6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore qframe qframe.
}
\examples{
test_data <- read.table(system.file("extdata", "test.tab", package = "FScanR"), header=TRUE, sep="\t")
FScanR(test_data)
}
\references{
1. X Chen, Y Jiang, F Gao, W Zheng, TJ Krock, NA Stover, C Lu, LA Katz & W Song (2019). 
Genome analyses of the new model protist Euplotes vannus focusing on genome rearrangement and resistance 
to environmental stressors. Molecular Ecology Resources, 19(5):1292-1308. 
<https://doi.org/10.1111/1755-0998.13023>
}
\author{
Xiao Chen
}
